src/job_aggregator/aggregator.py
"""
Job aggregator module for collecting job listings from various sources.
"""
import requests
from bs4 import BeautifulSoup

class JobAggregator:
    def __init__(self, config):
        self.config = config
        self.sources = {
            'linkedin': self._scrape_linkedin,
            'indeed': self._scrape_indeed,
            'stepstone': self._scrape_stepstone
        }
    
    def search_jobs(self, title, location, radius):
        """
        Search for jobs across all configured sources.
        """
        all_jobs = []
        
        for source_name, scraper_func in self.sources.items():
            try:
                jobs = scraper_func(title, location, radius)
                all_jobs.extend(jobs)
                print(f"Found {len(jobs)} jobs from {source_name}")
            except Exception as e:
                print(f"Error scraping {source_name}: {str(e)}")
        
        return all_jobs
    
    def _scrape_linkedin(self, title, location, radius):
        """
        Scrape jobs from LinkedIn.
        """
        # This is a placeholder implementation
        # In a real implementation, you would use LinkedIn's API or web scraping
        jobs = []
        # Example job structure
        jobs.append({
            'title': f'{title} at Company A',
            'company': 'Company A',
            'location': location,
            'url': 'https://linkedin.com/jobs/123',
            'description': 'Job description for Company A'
        })
        return jobs
    
    def _scrape_indeed(self, title, location, radius):
        """
        Scrape jobs from Indeed.
        """
        # This is a placeholder implementation
        jobs = []
        # Example job structure
        jobs.append({
            'title': f'Senior {title} at Company B',
            'company': 'Company B',
            'location': location,
            'url': 'https://indeed.com/jobs/456',
            'description': 'Job description for Company B'
        })
        return jobs
    
    def _scrape_stepstone(self, title, location, radius):
        """
        Scrape jobs from Stepstone.
        """
        # This is a placeholder implementation
        jobs = []
        # Example job structure
        jobs.append({
            'title': f'Lead {title} at Company C',
            'company': 'Company C',
            'location': location,
            'url': 'https://stepstone.com/jobs/789',
            'description': 'Job description for Company C'
        })
        return jobs